+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to
 
 ### Added
 
+- ✨(backend) add admin action to check domain health
 - ✨(dimail) check domain health
 - ✨(frontend) disable mailbox and allow to create pending mailbox
 - ✨(organizations) add siret to name conversion #584